refactor: create file handler functions for decompose and compose
