+++ src/commands/decomposer/decompose.ts
@@ -1,14 +1,11 @@
 /* eslint-disable */
-import * as fs from 'node:fs';
-import * as path from 'node:path';
-
 import { SfCommand, Flags } from '@salesforce/sf-plugins-core';
 import { Messages } from '@salesforce/core';
-import { INDENT, METADATA_DIR_DEFAULT_VALUE } from '../../helpers/constants.js';
+import { METADATA_DIR_DEFAULT_VALUE } from '../../helpers/constants.js';
 import jsonData from '../../metadata/metadata.js';
 import { Metadata } from '../../metadata/metadataInterface.js';
 import { getAttributesForMetadataType } from '../../service/getAttributesForMetadataType.js';
-import { xml2jsParser } from '../../service/xml2jsParser.js';
+import { decomposeFileHandler } from '../../service/decomposeFileHandler.js';
 
 Messages.importMessagesDirectoryFromMetaUrl(import.meta.url);
 const messages = Messages.loadMessages('sfdx-decomposer', 'decomposer.decompose');
@@ -48,7 +45,8 @@ export default class DecomposerDecompose extends SfCommand<DecomposerDecomposeRe
 
     if (metaAttributes) {
       const { metaSuffix, fieldNames, xmlElement, metadataPath } = metaAttributes;
-      this.parseMetadataFiles(metadataPath, metaSuffix, fieldNames, xmlElement);
+      decomposeFileHandler(metadataPath, metaSuffix, fieldNames, xmlElement);
+      this.log(`All metadata files have been decomposed for the metadata type: ${metaSuffix}`);
     } else {
       this.error(`Metadata type ${metadataTypeToRetrieve} not found.`);
     }
@@ -56,21 +54,5 @@ export default class DecomposerDecompose extends SfCommand<DecomposerDecomposeRe
     return {
       path: 'sfdx-decomposer\\src\\commands\\decomposer\\decompose.ts',
     };
-  }
-
-  private parseMetadataFiles(metadataPath: string, metaSuffix: string, fieldNames: string, xmlElement: string): void {
-    const files = fs.readdirSync(metadataPath);
-    files.forEach((file) => {
-      const filePath = path.join(metadataPath, file);
-      if (file.endsWith(`.${metaSuffix}-meta.xml`)) {
-        // Add your logic to parse the metadata file here
-        this.log(`Parsing metadata file: ${filePath}`);
-        const xmlContent = fs.readFileSync(filePath, 'utf-8');
-        const baseName = path.basename(filePath, `.${metaSuffix}-meta.xml`);
-        const outputPath = path.join(metadataPath, metaSuffix === 'labels' ? '' : baseName);
-        xml2jsParser(xmlContent, outputPath, fieldNames, xmlElement, baseName, metaSuffix, INDENT);
-      }
-    });
-    this.log(`All metadata files have been decomposed for the metadata type: ${metaSuffix}`);
   }
 }
